Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
openai==2.8.1
Werkzeug==3.0.1
email-validator==2.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2
xlrd==1.2.0
oletools==0.60.1

